fixed oxtype of default actions
in theory a zero value is invalid for oxtype, cause when creating a new action in the backend, it will have "1" for "action", "2" for "promotion" and "3" for "banner". So checking the type of the action inside own functions or templates often fails, cause some actions have oxtype "0" and some other have "1" although both are "action" actions.