bug fix #5842 Module settings default values are not shown before module activation.
Do not check whether default value for arr and aarr is an array, as it is checked in _arrayToMultiline and from metadata you can only pass arrays.
Adding handling of other settings default values, not only arr and aarr.
Adding handling for bool values to accept string, bool and int values ('true', true, 1, 'on'...).