directory rights check removed
it was redundand as system chec is already performed in first step by different code. It's not good to have duplicated code. It's not good to check in forth step as information wouldn't be handled correctly. This check ws wrong as it required directories exist even if shop can generate those directories. This will allow to pass setup without images needed for azure theme.