ESDEV-2624 Fix tests library
Move setting up of tests suite to be run at normal tests setup so that setUpBeforeClass method would not be static.
Including oxTestCase class from bootstrap_selenium so that modules would not need to include it themselves.
Checking if setupPath is actually set in _db.php so that passing empty path would not reset default path to search in.
Removing setting of default shopSetupPath as this causes problems when running tests on remote servers.
(cherry picked from commit fd649f0)