0004648: Moving articles invalidates old URL
Remove move to history part as moving to history part does not make sense at all. It seems that oxseohistory without corresponding oxseo record does not work, also oxseohistory hits are not recorded when oxseo record is marked as expired. In short: there is no sense in oxhistory at this point.