Revert Unify test results on different machines
Xml is still formed differently on different environments.

SimpleXML does not act in the same way on different machines with empty tags - it either generates <tag></tag> or </tag>.
These tags are both valid, but due to these differences tests are failing. To unify both during testing xml is converted to SimpleXML object and back to string.