Fixed access to target Category when loading products
Though a target category is given as function parameter, the "active category id" was loaded directly from Request in line 379. This did not make sense at all. The system does not behave as expected here, if the list controller is not used in standard (OXID-)case.

After some short tests, this change should not cause any problems.