Added support for PSR0 and PSR4 autoloading. This brings support for namespaces and makes it possible to include other 3rd party libaries to core.
This pull request was designed to be backward compatible, so i tried not to change old code as much as possible
to lower the risc of breaking bc breaks or new bugs or unmasking bugs.

PSR3 Logger Interface and a Logger Implementation will be part of an other pull request,
also not included in this merge request is PSRR3/PSR4 autoloading support for modules.

Performance:
- negative impact should be ver low (hopyfully not mesureable),
- positve impact depends on usage of namespaces.

I am also thinking of caching $aClassPaths in the autoloader to make performance improvements but this would also be an other merge request.