ESDEV-3028 oxSeoEncoderTag Fix _prepareTag method
Fix not working _prepareTag function and deprecate it, as this should have been done some time ago.
oxTagCloud class had deprecated method (prepareTag) which was used here. After deprecated method was removed,
this place was not updated. More so this method is not used anymore anywhere, so is deprecated from now on.