oxnew parameters are serialized and so they can hit servers memorylimit
this commit fixes a problem that happens if module-developers use oxNew with heavy constructor arguments (e.g. objects or arrays).
my modification checks the arguments first, only simple values (scalars/null) are accepted for cachable objects.

This commit does NOT fix the problem if someone uses a very big string as constructor argument,
but therefor strlen of the args could be checked additional.