Update oxseodecoder.php
In some cases you want to know, if the same URL exists in other shops. For example if you don't have a responsive shop and you want to have a link on your mobileshop which links to the same page in your desktopshop. This extension of the method could help with this case.