ESDEV-3181 0005928 Basket item amounts for personalisable articles
When changing a basket containing personalisable articles in a way the
personalisable articles now are exactly the same as one already in basket,
we need to make sure oxcmp_basket gets the overall amount of this article right.
Fix this in oxcmp_basket::_addItems. Add integration test that fails without the fix.