ESDEV-3219 Create ImportObjects with oxNew instead of new
Check for class_exists is now crucial and can not be removed. This is because currently oxNew
lowercases provided class and tries to autoload it that way. As file names now match case with class names,
it fails to autoload correct class. If class_exists is called with correct class name first, this class gets autoloaded
and no failures comes in next.

Check for class_exists can be removed after oxNew allows creation of namespaced classes.