ESDEV-3166 Change waitForAjax to assertElementText
Use assertElementText where element text can change when fully loaded (after ajax requests). It will wait for text to show up even if at the moment of checking this text does not appear to be correct.

This method was recently added to testing_library, so composer.lock is also updated.