oxscript must pack javascript into a string to execute it in a global scope, for that some characters must be replaced to generate a valid string.
In the old version it does only replace quotes but newlines are also problematic and must be escaped.

Without that fix you will get javascript errors in combination with varnish in some situations, please contact me if you need to reproduce that.

Implementation details: i used a extra method, which i copied from a asychron javascript support module that we are using in projects.