performance can be improved by using intervalls and REQUEST_TIME
by replacing the current time with a rounded request time it's possible for the querycache to cache the results of an sql query.

I used a 1 minutes interval for rounding so all queries can be cached for at least 1 minute, even that may not give a huge performance improvement for normal usecases,
in some situations (e.g high load on specific pages) it will be an performance improvement.

- removed duplicate code in oxarticle and call the parent method
this will lead to the fact that oxarticle also get the improved query

- added helperMethods in oxutildate class