performance can be improved by using intervalls and REQUEST_TIME
- use a 5 minutes intervall for active sql queries.
i am not sure if 5 minutes is to much but at least one minute intervalls
can be done without bc break and aditional efforts for making it configurable.

```
    //round up 1 minute so query cache can work
    $timestamp = ceil($timestamp/60)*60;

I did no performance test on this, would be great to get some feedback if it helps.

- removed duplicate code in oxarticle and call the parent method
this will lead to the fact that oxarticle also get the improved query

- added helperMethods in oxutildate class