seoencoder creates entries with the wrong objectid because the objectid is not refresh.
It happend on a database with hundreds of similar urls

same-title.html
same-title-oxid.html
same-title-oxid1.html
...

And i realized the expired flag was not updated(to 0) for an seourl of some articles.

From programming point of view oxobjectid bust be set in that sql statement because it is not part of the primary key.