ESDEV-3429 Adjust Smarty entry in composer
Providing version number with a caret '^' reveals the initial version
which was taken during the development process and thus shows the
safe version to revert to in case of problems. This also has the same
effect in terms of update, i.e. when used as 'x.y.*' will produce the
same end result as '^x.y.z'.