Add `testAllLanguageFilesForExistence`
This test checks if the defined files are available with
case-sensitive `file_exists` method. The check is OS and
file-system independent.

This also fixes the issue of having different case-sensitive option
for the Host file system, i.e. the given errors should be thrown
despite what file-system is in use.

Side-effect changes:
* Fix `_getLanguageFilePath` to be able to provide case sensitive
  path for language files inside the Setup directory
* Update language file entries which have case-sensitive changes,
  e.g. `en/setup` -> `En/Setup`