ESDEV-3462 Fix test bug and clean up test.
test bug: cause we didn't gave in a value for "select from master", the code assumed we wanted to fetch the values from the slave. The database class had a slave connection, with same credentials as the master, and fetched from there -> two connections to the same database makes transactions tests fail.

 clean up:
  - use assertSame instead assertEquals
  - comment some things for style guide