ESDEV-3635 Deprecate oxLegacyDb::query()
This method has the same signature and works internally exactly the same as oxLegacyDb::execute() with the only difference that it is not possible to enable logging for queries done with oxLegacyDb::query().
Replace all usages of oxLegacyDb::query() with oxLegacyDb::execute()
Remember to NOT use this function for select statements in a master-slave setup, as it will always execute the select query on the slave database.