ESDEV-3509 Reimplement LegacyDatabase functions: selectLimit()
Doctrine::selectLimit() has already been implemented before.
There was an  issue with the parameters not being sanitized. This issue was fixed. $offset and $rowCount are casted to int.
An E_USER_DEPRECATED error is triggered, if one of the arguments is not numeric.
Fixed tests.
Set user defined error handler in DatabaseInterfaceImplementationBaseTest::setUp() to be able to assert errors triggered with trigger_error().
(cherry picked from commit 1236d8d)