PR-580 Make adding template blocks more fast and reliable
background in automated deployment on heavy load it is important that we have helper methods that add new template blocks, with an minimum impact on the running system.

So the new code ensures that the method can be used to add template blocks without having to delete them first,
only outdated entries gets automatically deleted after new entries are added.