set html language regardless of performance settings
set the language attribute in html is even if performance option to load languages is not enabled.
To do so i use the shop internal language abbr, that should be avail without loading the setting from the database,
so we will still have performance benefit but without loosing functionality to set html language
(in most cases, in special if internal language abbr matches the ISO 639-1 language abbr).