this change prevents the shop from disabling modules or going into an unrecoverable state if modules are broken
we already have that blDoNotDisableModuleOnError configuration option, that give us some control how the shop should behave on module errors,
but in projects it has shown, that the current behavior is not a good default no matter how you set this variable.

If you set
blDoNotDisableModuleOnError to true, the shop may disable modules that you need only because of an temporary error.
and if you set blDoNotDisableModuleOnError to false, an module error may bring the shop in an unrecoverable state where can not fix the module anymore.

so i introduced a new default that ignores errors, and tries to continue. You will still see the exception in logfiles, and maybe the broken module will may stop some areas of the shop from working,
but it's likely that will be able to fix the problem by module internals or by activating/deactivating the module in the backend.

use singelton oxutilsobject also in unittests to preserve internal state