removing parents from view chain after processing
Remove all parent views that where added to the chain before

How it was wrong before the change?

The issue comes when the view chain is modified, then the shop adds a new parent to the chain (which is correct), but after processing the controller (normally widgets), that parent is not removed from the chain. then we get strange errors like method not found, or sessions that are not saved. This kind of error is very hard to debug and to understand, but unlikely to happen because we do not often use parents in view chains.

How it will be better with the change?

With that change we are going to cleanup afterward, so every parent that gets added to the view chain will be removed after processing the widget.