not forcing session on POST requests
Forcing a php session for http post requests makes trouble in several projects, because sessions may created on ajax request or form submits.
It is not always true that an POST request will need a session, the only thing "POST" does is storing data on the server and this can be independent from the current session.

On the other hand, forcing the session on POST requests may simplify other places, because there the code does not explicitly have to start the session. For that reason i introduced and automatic session start if calling code tries to variables in the session.

 this pull request should be tested with care it is still experimental.