ESDEV-4007 PR #432 Fix bug 0006407
The Language abbreviation that was saved with a '-' gets replaced with '_' when URI is created.
But the comparison made in oxseoencoder.php, compares the uri with the Language abbreviation from the database.
Hence do not match. And so repeatedly added to the URI for Languages with abbreviation using an Underscore.