Using Registry until we have DI
Using new should be avoided because it allocates new memory,
even if oxNew would clone the object (with cow) support it is sometimes better to reuse objects.
For the Request object this is the case because someday you might want to give it internal state, it may need to parse post data from input stream once, or you may like to manipulate that data within the control-flow of the application.

Closes #459.