ESDEV-4008 #397 Backport Fix nopic generation for png and gif
Bug ID 0006291

imagecreatefrompng was used to create the target image in the method OxPicGenerator::resizePng() without considering the source image format. Hence when a false image was requested in any format other than jpg(since nopic file currently present in jpg), no image gets printed. With this solution, if an unavailable png file is requested, a nopic.png is generated and output. The same goes for GIF images too.