Revert "ESDEV-4004 0005307 PR #460 Fix checkParamSpecialChars to not modify reference"
This bugfix introduces a BC break and must be reverted.
This bug cannot be fixed in v5.3
The right way to fix this bug in v6.0 is to create a separate method in v5.3 and deprecate the current method checkParamSpecialChars().
In v6.0 do not change the current method, but replace it by the new one and delete the current method checkParamSpecialChars()

This reverts commit 3eec83cf8c9f8ac4d3faad627d0200acfd1aab32.