ESDEV-4385 Fix BasketFrontendTest
Under some specific and unique environment conditions such as:

* PHP setting `request_order` is equal to `CGP` or empty value
* Testing library is using `is_subshop` variable as `true`
* Testing library is using `activate_all_modules` as `true`

The test case `BasketFrontendTest::testFrontendOrderStep4ChangedAddress`
Fails because it tries to re-use a basket which was created on a
different subshop. Different subshop has different activated modules
which causes the `Basket` object to be different for different
subshops and different `Basket` objects are not compatible with
each other thus leading for the test to fail.

The fix tries to solve the issue by ensuring that additional windows
are opened within the same subshop (context) which leads to having
same (and compatible) `Basket` objects.