ESDEV-3779 Provide user for getPaymentList
task: ESDEV-4571

`getPaymentList` implementation inside the shop is able to handle
the case when no user is specified, unfortunately not all modules
are created with this same expectation thus this change allows
to not cause fatal errors for modules which are not capable of
working with undefined users.

This issue was noticed while working with PayOne module.