ESDEV-3779 Split test to two parts
It makes sense to check if correct error is returned even when modules are active.
It makes no sense to check if methods are called or not called when module is active.

For example method getPayment is called by PayOne to check
if Payment is managed by module. If not original functionality applies.
This makes test fails as it expects method to NOT be called.