ESDEV-4546 Fix failing test
Using instances of concrete classes is not expected to work when functions
use type hints from the unified namespace. PHP 5.6 and PHP >= 7 handle this
error in a different manner. The test is not skipped any longer, but deals
with this fact.

Changed file names of tests to according the classes they contain