Fix: E_USER_DEPRECATED errors will not trigger offline page
If error_get_last() would return an error of type E_USER_DEPRECATED an
offline page was rendered.
E_USER_DEPRECATED are not the kind of errors, which should trigger this
behavior, as module or component writers should have the freedom to notify
about deprecated code without expecting the shop to go offline.