Update oxseoencoder.php
i think, there was a minor bug in oxseoencoder, which should be fixed with this.. SEO-URLs are stored separately for every subshop. In oxseoencoder->_loadFromDb() SEO-Urls, which are marked as expired, were copied to seohistory-Table for given ObjectID and shopID, but the expired-Flag was unset for given ObjectID and every Subshop (given shopId wasnt used). This should work better now.