ESDEV-4827 Add unified namespace compatible version of ide-helper
The ide-helper is needed in order to use features like autocompletion for backwards compatible classes (e.g. class oxArticle) in your ide. The unified namespace classes are real classes and therefor recognized by any IDE. But for backwards compatible classes only a class_alias is created. Not every IDE recognizes the usage of the PHP method class_alias. E.g. PhpStorm knows how to deal with it from version 2016.3 on.