OXDEV-470 Remove market selection from Backend
Some places, where the parameter sShopCountry was used were overlooks.
This commit fixes some usages and tests.
The "country"-part of \OxidEsales\EshopCommunity\Application\Controller\Admin\AdminController::getServiceUrl() is now hardcoded to "international", which is the default value for \OxidEsales\EshopCommunity\Application\Controller\Admin\AdminController::_getCountryByCode() if an empty parameter is passed.
The return value of AdminController::getServiceUrl() was used for the dynPage functionality, which has been disabled and deprecated.