PR-634 Reuse _getUrlExtension
The hard-coded value is the same which is already provided by `_getUrlExtension`.
To me it also looks that not only the value but the intent of method is also the same
and matches this case.

This change would let me reduce the number of methods I would need to override for
a project, i.e. only `_getUrlExtension` would be enough.

Related #634