OXDEV-992 Move methods to FrontendController
The following methods will be moved to FrontendController
- isUserAllowedToManageOwnReviews
- getReviewAndRatingItemsCount
- getContainer

Some module writers include "others controllers" templates in their
templates in order to use them as a kind of cheap kind of widget.
This is not how templates meant to be used, as the separation of concerns
between the controllers is no longer guaranteed.
For the sake of better compatibility with these modules, the methods
mentioned above are included in the FrontendController.