PR-661 Deal with broken modules
With this commit the Oxid shop is able to handle broken modules in a proper way, by logging the exception without trying to disable the module. The old configurable behavior was eighter to throw a exception and by that stopping the shop from working or to disable the module, which could lead to endless loop or to unwanted behavior.

- Fix test
- Assert that exception is logged
- avoid double errors