PR-663 Fix unusable shop after activation of a module with migrated metadata (v2)
When a module that was previously activated with metadata v1 is activated after its metadata has been migrated to v2, the old extensions are now correctly purged from the database.

Additionally, the ModuleChainsGenerator now correctly handles special cases where two modules have extensions to oxconfig and one of them is not available anymore (e.g. because of migrated metadata).