PR-683 Ensure temp file in tmp directory
with this fix temporary files are ensured to be created within the tmp directory
the issue without the fix is that in a secure envuornment and without a trailing slash in compiledir folder setting  the confic.inc.php the temp file can not be created and for example orders failed.
a workaround for projects is to lower security for writing next to the tmp folder or to add a trailing slash in the setting for the tmp folder.
no bugtracker entry created yet by me