OXDEV-1820 Refactor PasswordHash services to check PHP warnings
The documentation of password_hash states, that the return value is a
string (on success) or false (on error).
Sadly this is not true as also NULL can be returned in some edge cases.
As the warnings, which are issued in such cases contain valuable
information about the reason of the failure, the WARNING message is added
to the exception message