PR-698 Read config parameter by getConfigParam not getRequestParam
By calling ->getRequestParam('blVariantSelection') we will always get null, as this request is never made and leads to this bug: https://bugs.oxid-esales.com/view.php?id=6968
I replaced the call with the method getConfigParam, like it is in the other Ajax methods done. After that the variants are shown in the manufactor popup list, too.