OXDEV-2279 Make deprecation comment more clear
The deprecated methods still work in terms of "doing the same thing as before": e.g. hash a string with MD5. It should be clear that "doing the same thing as before" might not be sufficient any more to hash the password or login a user. The current password hash might be already updated and not compatible with the old method any more.