Make Backend "Save in Language" compatible with lowercase and CamelCase fields
Previously oxid was not able to provide the "save in lang" dropdown in the backend of custom models if these had lowercase or camelcase fieldnames (e.g. mytable__title instead of mytable__TITLE). Since all of the shop is fine with lowercase fields, it makes sense to remove that final barrier.