Pass optional cookie parameters to setSessionCookie
Adds the optional parameters used by Registry::getUtilsServer()->setOxCookie(). For example, in some applications it is necessary that the expire time is set to a value and not the default of 0, so that the session can be controlled to live longer than the current browser session.

Adds a standard function comment header.