Remove wrong code for getting module URL from admin
Removing code that assumes that the module URL has anything to do with the admin URL. Probably even the rest of the code can also be simplified a lot.
The code is not just unnecessary, it also causes problems when the shop admin runs under a completely different URL (e.g. a subdomain) than the rest of the shop. (Which is often the case for bigger customers).