Improve redirect cycle detection. This fixes issues when redirect cascade is happening.
## How to Reproduce:

The old code successfully stops the browser from running in a cycle, but it is not able to detect if it is really a cylce or just a second redirect after the first one.
In the project where I hit this limitation there was a simple redirect from basket to startpage if the user is not logged in.
As the normal logout function (in ssl mode) does a redirect, a logout on the basket page caused a sequenze of two redirects.

## Redirect Counting

Counting the redirect does not really detect a cycle but is good enough to detect problems. The hardcoded 3 is arguable some browser may allow 16 or more.
Having a serverside cycle detection is better to be able to provide better error messages and to be able monitor the application for errors on server side, the value should therefore be lower then 16.
Also multiple redirects should be avoided to have good user experience so value lower then 4 is preferable but also you do not want to optimize big parts of the code base so 1-3 redirects may be acceptable
see
https://stackoverflow.com/questions/10895406/what-is-the-maximum-number-of-http-redirections-allowed-by-all-major-browsers

## Use of Exception Instead of Return

The old code used a simple return statement, so the caller code could continue handling the edgecase but after research I found no caller code in projects and module that was able to handle this return correctly. often the caller code assumed that the redirect function will never return, so the behavior was causing following errors that were hard to understand.
It also not a good alternative to exit the code if a cycle is detected because the user needs some feedback in the browser to understand that a error did happen and to navigate back to start page or to report the issue.
Throwing a exceptions allows the caller code to deal with the exception and also to framework to take actions in that situation (e.g. showing the offline page and logging the error) which helps developers to understand and handle the situation.