PR-869 Fix cached delivery rules reuse
The bug occurs when having the B2B modules activated. When increasing the amount of a basket position in the basket overview so that a delivery rule which has zero cost
should be pulled, the old cost is still visible and only updated when pushing the update button a second time.
Initializing the member variable _aDeliveries to an emty array fixes this behaviour.

Related #869