Remove redirects to directory
why:
it can make trouble if apache adds internal ports to the redirect, it makes oxid not flexible to configure seo url by modules (e.g. you may not like to have the slash or .html extension and it does not add value because seo.php can also handle the redirect so it is useless code.