using OXUSERNAME in addRegister() instead of OXID
The problem is: The register event fires the user's OXID via addRegister(). In the Emos.php within the addRegister() method the user's OXID is hashed again: $this->_registerUser = md5($sUserId);
So econda would receive a md5(OXUSERNAME) from the addLogin(), but unfortunately a different hash, a md5(OXID), which would be a md5(md5(OXUSERNAME)).

So the (bug?) fix could be this one, so that both methods the addRegister() and the addLogin() receive the OXUSERNAME. but return the hashed OXUSERNAME: md5(OXUSERNAME).

Or (second approach): You could modify the Emos.php and remove the md5() within the addRegister() method.

With both suggestions econda would receive the same hash for both the register and the login event.